Refactor search screen associated files
